Rework project
Reduced code and made Nim code generation use std/macros NimNode instead
of writing Nim code as strings.

Compared to before, ADTs (variant objects; objects with
`case`-statements) are not generated anymore. Motivation is for easier
marshalling. XSD doesn't define a discriminator so it's more complex to
find the correct case-branch to set in the object.